Create separate files for each component in the home page
